feat: Abstract the `get` helper
